Merging latest dev changes into new release branch off main

+++ ArchiveADataset.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
+source $(pwd)/.env
+if [[ -z $DATASETS_DIR ]] || [[ -z $IMAGE ]]; then
+  echo "Error: DATASETS_DIR and/or IMAGE not set in docker compose .env file. Exiting."
+  exit 1
+fi
 docker run --rm -it \
   -v "${DATASETS_DIR}:/datasets" \
   -v "$(pwd)/logs:/erddapData/logs" \
   -v "$(pwd)/erddap/content:/usr/local/tomcat/content/erddap" \
-  -v "$(pwd)/erddap/data:/erddapData" \                                                                                                                                                                                                                                          axiom/docker-erddap:latest \
+  -v "$(pwd)/erddap/data:/erddapData" \
+  ${IMAGE} \
   bash -c "cd webapps/erddap/WEB-INF/ && bash ArchiveADataset.sh -verbose"
